requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.36.0
django-cors-headers==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.3
drf-dynamic-fields==0.3.1
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==6.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.8.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.24.3
whitenoise==4.1.2
